Added Assert GGP.
git-svn-id: https://gforge.sci.utah.edu/svn/MRLProjects/Plugins/AngioFE2@9687 b505bdfe-9d5c-0410-8600-697d5fc8e487

+++ AngioFE2/Fileout.cpp
@@ -73,7 +73,7 @@ void Fileout::printStatus(FEAngio& angio)
 // Save microvessel position at the current time point
 void Fileout::save_vessel_state(FEAngio& angio)
 {
-	unsigned long int segcount = 0;
+	unsigned int segcount = 0;
 	for (size_t i = 0; i < angio.m_pmat.size(); i++)
 	{
 		Culture* cult = angio.m_pmat[i]->m_cult;
